@@ -9,17 +9,17 @@
     "@matrix-widget-toolkit/api": "3.3.1",
     "@matrix-widget-toolkit/react": "1.0.6",
     "@matrix-widget-toolkit/semantic-ui": "1.0.8",
-    "@reduxjs/toolkit": "^1.9.7",
+    "@reduxjs/toolkit": "^2.2.3",
     "i18next": "^23.11.2",
     "i18next-chained-backend": "^4.6.2",
     "i18next-http-backend": "^2.5.1",
     "joi": "^17.13.0",
     "lodash": "^4.17.21",
     "matrix-widget-api": "^1.6.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-i18next": "^14.1.1",
-    "react-redux": "^8.1.3",
+    "react-redux": "^9.1.0",
     "react-router-dom": "^6.23.0",
     "react-scripts": "5.0.1",
     "react-use": "^17.5.0",
@@ -55,12 +55,12 @@
     "@matrix-widget-toolkit/testing": "2.4.0",
     "@semantic-ui-react/craco-less": "^3.0.0",
     "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^14.2.2",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^27.5.2",
     "@types/node": "^20.8.6",
-    "@types/react": "^17.0.45",
-    "@types/react-dom": "^17.0.16",
+    "@types/react": "^18.2.74",
+    "@types/react-dom": "^18.2.24",
     "i18next-parser": "^8.13.0",
     "msw": "^2.2.14",
     "typescript": "^5.4.5"
