@@ -10,17 +10,17 @@
   "devDependencies": {
     "@craco/craco": "^7.1.0",
     "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^14.2.2",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^27.5.2",
     "@types/jest-axe": "^3.5.9",
     "@types/lodash": "^4.17.0",
     "@types/node": "^20.8.6",
-    "@types/react": "^17.0.45",
+    "@types/react": "^18.2.74",
     "copyfiles": "^2.4.1",
     "i18next-parser": "^8.13.0",
     "jest-axe": "^8.0.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-scripts": "5.0.1",
     "typescript": "^5.4.5"
   },
@@ -52,7 +52,7 @@
     "i18next-resources-to-backend": "^1.2.1",
     "lodash": "^4.17.21",
     "matrix-widget-api": "^1.6.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-i18next": "^14.1.1",
     "react-use": "^17.5.0"
   },
