@@ -12,12 +12,12 @@
     "@craco/craco": "^7.1.0",
     "@semantic-ui-react/craco-less": "^3.0.0",
     "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^14.2.2",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^27.5.2",
     "@types/node": "^20.8.6",
-    "@types/react": "^17.0.45",
-    "react": "^17.0.2",
+    "@types/react": "^18.2.74",
+    "react": "^18.2.0",
     "copyfiles": "^2.4.1",
     "i18next-parser": "^8.13.0",
     "react-scripts": "5.0.1",
@@ -44,7 +44,7 @@
     "i18next-browser-languagedetector": "^7.2.1",
     "i18next-resources-to-backend": "^1.2.1",
     "matrix-widget-api": "^1.6.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-i18next": "^14.1.1",
     "react-use": "^17.5.0",
     "semantic-ui-less": "^2.5.0",
