@@ -36,10 +36,9 @@ import {
   Stack,
   TextField,
 } from '@mui/material';
-import { unstable_useId as useId } from '@mui/utils';
 import { isError } from 'lodash';
 import { MatrixCapabilities } from 'matrix-widget-api';
-import { ReactElement, useEffect, useState } from 'react';
+import { ReactElement, useEffect, useId, useState } from 'react';
 import { NavigationBar } from '../NavigationPage';
 
 /**
