@@ -54,7 +54,8 @@ describe('<IdentityPage />', () => {
     ).toBeInTheDocument();
 
     await expect(screen.findByText(/Identity/i)).resolves.toBeInTheDocument();
-    expect(screen.getByText('Error')).toBeInTheDocument();
+
+    expect(await screen.findByText('Error')).toBeInTheDocument();
     expect(screen.getByText(/error: .+/i)).toBeInTheDocument();
   });
 
